Moved FromBitmap static method to separate Factory class. Fixes #83
